requirements.txt
streamlit
pandas
plotly
openpyxl
fpdf2
openai
fuzzywuzzy
python-Levenshtein
matplotlib
statsmodels